fix(...) : 약속만들기 쪽 구조 변경 - activity 안쓰기
