+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
-        <activity android:name=".presentation.make_meet.MeetRegisterActivity"
+        <activity android:name=".presentation.make_meet.MakeMeetActivity"
             android:windowSoftInputMode="adjustResize"/>
 
         <activity
