+++ app/src/main/java/com/manna/HomeActivity.kt
@@ -18,7 +18,7 @@ import androidx.fragment.app.commitNow
 import com.manna.common.BaseActivity
 import com.manna.databinding.ActivityHomeBinding
 import com.manna.databinding.DialogWelcomeBinding
-import com.manna.presentation.make_meet.MeetRegisterActivity
+import com.manna.presentation.make_meet.MakeMeetActivity
 import com.manna.util.ViewUtil
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -77,7 +77,7 @@ class HomeActivity : BaseActivity<ActivityHomeBinding>(R.layout.activity_home) {
         changeTab(0)
         binding.run {
             addButton.setOnClickListener {
-                startActivity(MeetRegisterActivity.getIntent(this@HomeActivity))
+                startActivity(MakeMeetActivity.getIntent(this@HomeActivity))
             }
 
             meetListTab.setOnClickListener {
